requirements.txt
Flask==3.0.0
Pillow==12.0.0
Werkzeug==3.0.1

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
